refactor: isolate symbolic rotation fusion
Move fusion policy, GF(2) preparation, matrix composition, and scalar execution into a dedicated sampling module. Add focused coverage for dynamic signs, signed identity phases, short runs, and rank-three fallback.
