+++ src/clifft/sampling/fused_rotation.cc
@@ -1,0 +1,301 @@
+#include "clifft/sampling/fused_rotation.h"
+
+#include "clifft/sampling/kernels.h"
+
+#include <algorithm>
+#include <array>
+#include <bit>
+#include <cassert>
+#include <complex>
+
+namespace clifft::sampling {
+
+namespace {
+
+// A general matrix sweep costs more than one Pauli butterfly, so match the
+// legacy fusion threshold and require it to replace at least three sweeps.
+constexpr size_t kMinFusedRotationActions = 3;
+
+// Each independent representative parity doubles the matrix table. Five
+// covers the measured QV workloads while bounding a descriptor at 32 matrices.
+constexpr uint32_t kMaxFusedRotationSelectors = 5;
+
+class BinaryBasis {
+  public:
+    [[nodiscard]] bool insert(uint64_t value, uint32_t max_rank = 64) {
+        uint64_t reduced = value;
+        for (int pivot = 63; pivot >= 0; --pivot) {
+            const uint64_t bit = uint64_t{1} << pivot;
+            if ((reduced & bit) != 0 && rows_[pivot] != 0) {
+                reduced ^= rows_[pivot];
+            }
+        }
+        if (reduced == 0) {
+            return true;
+        }
+        if (rank_ == max_rank) {
+            return false;
+        }
+        const uint32_t pivot = 63U - static_cast<uint32_t>(std::countl_zero(reduced));
+        const uint64_t bit = uint64_t{1} << pivot;
+        for (uint64_t& row : rows_) {
+            if ((row & bit) != 0) {
+                row ^= reduced;
+            }
+        }
+        rows_[pivot] = reduced;
+        ++rank_;
+        return true;
+    }
+
+    [[nodiscard]] uint32_t rank() const { return rank_; }
+
+    [[nodiscard]] std::vector<uint64_t> rows() const {
+        std::vector<uint64_t> result;
+        result.reserve(rank_);
+        for (uint64_t row : rows_) {
+            if (row != 0) {
+                result.push_back(row);
+            }
+        }
+        return result;
+    }
+
+  private:
+    std::array<uint64_t, 64> rows_{};
+    uint32_t rank_ = 0;
+};
+
+uint32_t basis_coordinates(uint64_t value, std::span<const uint64_t> rows) {
+    uint32_t coordinates = 0;
+    for (size_t i = 0; i < rows.size(); ++i) {
+        const uint64_t pivot = std::bit_floor(rows[i]);
+        if ((value & pivot) != 0) {
+            value ^= rows[i];
+            coordinates |= uint32_t{1} << i;
+        }
+    }
+    assert(value == 0 && "value must belong to the prepared binary span");
+    return coordinates;
+}
+
+void multiply_matrix_left(std::span<std::complex<double>> matrix,
+                          std::span<const std::complex<double>> left, size_t dimension) {
+    std::array<std::complex<double>, 16> result{};
+    for (size_t row = 0; row < dimension; ++row) {
+        for (size_t column = 0; column < dimension; ++column) {
+            for (size_t inner = 0; inner < dimension; ++inner) {
+                result[row * dimension + column] +=
+                    left[row * dimension + inner] * matrix[inner * dimension + column];
+            }
+        }
+    }
+    std::copy_n(result.begin(), dimension * dimension, matrix.begin());
+}
+
+std::optional<PreparedFusedRotation> prepare_fused_rotation(std::span<const PlannedAction> actions,
+                                                            const BinaryBasis& orbit_basis) {
+    assert(actions.size() >= kMinFusedRotationActions &&
+           "fused rotation requires at least three actions");
+    const uint32_t active_width = actions.front().active_before;
+    const std::vector<uint64_t> orbit_rows = orbit_basis.rows();
+    uint64_t orbit_pivots = 0;
+    for (uint64_t row : orbit_rows) {
+        orbit_pivots |= std::bit_floor(row);
+    }
+
+    BinaryBasis selector_basis;
+    for (const PlannedAction& planned : actions) {
+        const auto& rotation = std::get<RotateActivePauli>(planned.action);
+        assert(planned.active_before == active_width && planned.active_after == active_width &&
+               rotation.sign.terms().empty() && !rotation.pauli.is_identity() &&
+               "fused rotation input must be a constant-sign nonidentity run");
+        if (!selector_basis.insert(rotation.pauli.z & ~orbit_pivots, kMaxFusedRotationSelectors)) {
+            return std::nullopt;
+        }
+    }
+    const std::vector<uint64_t> selector_rows = selector_basis.rows();
+
+    PreparedFusedRotation result;
+    result.active_width = active_width;
+    result.orbit_rank = orbit_basis.rank();
+    for (size_t i = 0; i < orbit_rows.size(); ++i) {
+        result.orbit_masks[i] = orbit_rows[i];
+        result.orbit_pivots[i] =
+            static_cast<uint32_t>(std::countr_zero(std::bit_floor(orbit_rows[i])));
+    }
+    result.selector_masks = selector_rows;
+
+    const size_t dimension = size_t{1} << result.orbit_rank;
+    const size_t matrix_size = dimension * dimension;
+    const size_t num_variants = size_t{1} << selector_rows.size();
+    result.matrices.resize(num_variants * matrix_size);
+    for (size_t variant = 0; variant < num_variants; ++variant) {
+        std::span<std::complex<double>> matrix(result.matrices.data() + variant * matrix_size,
+                                               matrix_size);
+        for (size_t diagonal = 0; diagonal < dimension; ++diagonal) {
+            matrix[diagonal * dimension + diagonal] = 1.0;
+        }
+
+        for (const PlannedAction& planned : actions) {
+            const auto& rotation = std::get<RotateActivePauli>(planned.action);
+            const PreparedRotation prepared =
+                prepare_rotation(rotation.pauli, active_width, rotation.half_turns);
+            const uint32_t x_coordinates = basis_coordinates(rotation.pauli.x, orbit_rows);
+            const uint32_t selector_coordinates =
+                basis_coordinates(rotation.pauli.z & ~orbit_pivots, selector_rows);
+            uint32_t local_z = 0;
+            for (size_t i = 0; i < orbit_rows.size(); ++i) {
+                local_z |=
+                    static_cast<uint32_t>(std::popcount(orbit_rows[i] & rotation.pauli.z) & 1U)
+                    << i;
+            }
+            const bool representative_phase =
+                (std::popcount(static_cast<uint32_t>(variant) & selector_coordinates) & 1U) != 0;
+            const double sine = rotation.sign.constant() ? -prepared.sine : prepared.sine;
+
+            std::array<std::complex<double>, 16> unitary{};
+            for (size_t column = 0; column < dimension; ++column) {
+                unitary[column * dimension + column] += prepared.cosine;
+                const bool odd_phase =
+                    representative_phase !=
+                    ((std::popcount(static_cast<uint32_t>(column) & local_z) & 1U) != 0);
+                const std::complex<double> phase =
+                    odd_phase ? -prepared.pauli.even_phase : prepared.pauli.even_phase;
+                unitary[(column ^ x_coordinates) * dimension + column] +=
+                    std::complex<double>{0.0, -sine} * phase;
+            }
+            multiply_matrix_left(matrix,
+                                 std::span<const std::complex<double>>(unitary).first(matrix_size),
+                                 dimension);
+        }
+    }
+    return result;
+}
+
+uint64_t insert_zero_bit(uint64_t packed, uint32_t pivot) {
+    const uint64_t lower_mask = (uint64_t{1} << pivot) - 1;
+    return (packed & lower_mask) | ((packed & ~lower_mask) << 1);
+}
+
+template <size_t Dimension>
+void apply_fused_rotation_orbits(State& state, const PreparedFusedRotation& rotation) noexcept {
+    static_assert(Dimension == 1 || Dimension == 2 || Dimension == 4);
+    const uint64_t orbit_count = state.size() / Dimension;
+    const size_t matrix_size = Dimension * Dimension;
+    assert(rotation.matrices.size() ==
+               (size_t{1} << rotation.selector_masks.size()) * matrix_size &&
+           "fused rotation matrix table must cover every selector value");
+
+    double* const real = state.real_data();
+    double* const imag = state.imag_data();
+    for (uint64_t packed = 0; packed < orbit_count; ++packed) {
+        uint64_t representative = packed;
+        if constexpr (Dimension >= 2) {
+            representative = insert_zero_bit(representative, rotation.orbit_pivots[0]);
+        }
+        if constexpr (Dimension == 4) {
+            representative = insert_zero_bit(representative, rotation.orbit_pivots[1]);
+        }
+
+        size_t selector = 0;
+        for (size_t bit = 0; bit < rotation.selector_masks.size(); ++bit) {
+            selector |= static_cast<size_t>(
+                            std::popcount(representative & rotation.selector_masks[bit]) & 1U)
+                        << bit;
+        }
+        const std::complex<double>* const matrix =
+            rotation.matrices.data() + selector * matrix_size;
+
+        std::array<uint64_t, Dimension> indices{};
+        std::array<double, Dimension> input_real{};
+        std::array<double, Dimension> input_imag{};
+        for (size_t column = 0; column < Dimension; ++column) {
+            uint64_t index = representative;
+            if constexpr (Dimension >= 2) {
+                if ((column & 1U) != 0) {
+                    index ^= rotation.orbit_masks[0];
+                }
+            }
+            if constexpr (Dimension == 4) {
+                if ((column & 2U) != 0) {
+                    index ^= rotation.orbit_masks[1];
+                }
+            }
+            indices[column] = index;
+            input_real[column] = real[index];
+            input_imag[column] = imag[index];
+        }
+
+        for (size_t row = 0; row < Dimension; ++row) {
+            double output_real = 0.0;
+            double output_imag = 0.0;
+            for (size_t column = 0; column < Dimension; ++column) {
+                const std::complex<double> weight = matrix[row * Dimension + column];
+                output_real +=
+                    weight.real() * input_real[column] - weight.imag() * input_imag[column];
+                output_imag +=
+                    weight.real() * input_imag[column] + weight.imag() * input_real[column];
+            }
+            real[indices[row]] = output_real;
+            imag[indices[row]] = output_imag;
+        }
+    }
+}
+
+}  // namespace
+
+FusedRotationRun prepare_fused_rotation_run(std::span<const PlannedAction> actions) {
+    FusedRotationRun result;
+    if (actions.empty()) {
+        return result;
+    }
+    const PlannedAction& first = actions.front();
+    const auto* first_rotation = std::get_if<RotateActivePauli>(&first.action);
+    if (first_rotation == nullptr || !first_rotation->sign.terms().empty() ||
+        first_rotation->pauli.is_identity()) {
+        return result;
+    }
+
+    BinaryBasis orbit_basis;
+    while (result.action_count < actions.size()) {
+        const PlannedAction& candidate = actions[result.action_count];
+        const auto* rotation = std::get_if<RotateActivePauli>(&candidate.action);
+        if (rotation == nullptr || !rotation->sign.terms().empty() ||
+            rotation->pauli.is_identity() || candidate.active_before != first.active_before) {
+            break;
+        }
+        BinaryBasis next_basis = orbit_basis;
+        if (!next_basis.insert(rotation->pauli.x, 2)) {
+            break;
+        }
+        orbit_basis = next_basis;
+        ++result.action_count;
+    }
+
+    if (result.action_count >= kMinFusedRotationActions) {
+        result.rotation = prepare_fused_rotation(actions.first(result.action_count), orbit_basis);
+    }
+    return result;
+}
+
+void apply_fused_rotation(State& state, const PreparedFusedRotation& rotation) noexcept {
+    assert(state.active_width() == rotation.active_width &&
+           "fused rotation width must match the active state");
+    switch (rotation.orbit_rank) {
+        case 0:
+            apply_fused_rotation_orbits<1>(state, rotation);
+            return;
+        case 1:
+            apply_fused_rotation_orbits<2>(state, rotation);
+            return;
+        case 2:
+            apply_fused_rotation_orbits<4>(state, rotation);
+            return;
+        default:
+            assert(false && "fused rotation orbit rank must be at most two");
+            return;
+    }
+}
+
+}  // namespace clifft::sampling
