+++ src/clifft/sampling/fused_rotation.h
@@ -1,0 +1,41 @@
+#pragma once
+
+#include "clifft/sampling/plan.h"
+#include "clifft/sampling/state.h"
+
+#include <array>
+#include <complex>
+#include <cstddef>
+#include <cstdint>
+#include <optional>
+#include <span>
+#include <vector>
+
+namespace clifft::sampling {
+
+// Applies one precomposed unitary to every orbit generated by at most two X
+// masks. Matrices are selected by representative-only Z parities, so each
+// coefficient is still loaded and stored exactly once.
+struct PreparedFusedRotation {
+    uint32_t active_width = 0;
+    uint32_t orbit_rank = 0;
+    // The orbit and selector masks are reduced GF(2) bases in ascending pivot
+    // order; orbit_pivots identifies the bits omitted from representatives.
+    std::array<uint64_t, 2> orbit_masks{};
+    std::array<uint32_t, 2> orbit_pivots{};
+    std::vector<uint64_t> selector_masks;
+    std::vector<std::complex<double>> matrices;
+};
+
+// Describes the maximal rank-two-eligible constant-sign rotation run beginning
+// at the supplied action. A populated rotation replaces all action_count
+// inputs; otherwise the caller lowers that many actions individually.
+struct FusedRotationRun {
+    size_t action_count = 0;
+    std::optional<PreparedFusedRotation> rotation;
+};
+
+[[nodiscard]] FusedRotationRun prepare_fused_rotation_run(std::span<const PlannedAction> actions);
+void apply_fused_rotation(State& state, const PreparedFusedRotation& rotation) noexcept;
+
+}  // namespace clifft::sampling
