+++ src/clifft/sampling/kernels.cc
@@ -78,71 +78,6 @@ uint64_t insert_zero_bit(uint64_t packed, uint32_t pivot) {
     return (packed & lower_mask) | ((packed & ~lower_mask) << 1);
 }
 
-template <size_t Dimension>
-void apply_fused_rotation_orbits(State& state, const PreparedFusedRotation& rotation) noexcept {
-    static_assert(Dimension == 1 || Dimension == 2 || Dimension == 4);
-    const uint64_t orbit_count = state.size() / Dimension;
-    const size_t matrix_size = Dimension * Dimension;
-    assert(rotation.matrices.size() ==
-               (size_t{1} << rotation.selector_masks.size()) * matrix_size &&
-           "fused rotation matrix table must cover every selector value");
-
-    double* const real = state.real_data();
-    double* const imag = state.imag_data();
-    for (uint64_t packed = 0; packed < orbit_count; ++packed) {
-        uint64_t representative = packed;
-        if constexpr (Dimension >= 2) {
-            representative = insert_zero_bit(representative, rotation.orbit_pivots[0]);
-        }
-        if constexpr (Dimension == 4) {
-            representative = insert_zero_bit(representative, rotation.orbit_pivots[1]);
-        }
-
-        size_t selector = 0;
-        for (size_t bit = 0; bit < rotation.selector_masks.size(); ++bit) {
-            selector |= static_cast<size_t>(
-                            std::popcount(representative & rotation.selector_masks[bit]) & 1U)
-                        << bit;
-        }
-        const std::complex<double>* const matrix =
-            rotation.matrices.data() + selector * matrix_size;
-
-        std::array<uint64_t, Dimension> indices{};
-        std::array<double, Dimension> input_real{};
-        std::array<double, Dimension> input_imag{};
-        for (size_t column = 0; column < Dimension; ++column) {
-            uint64_t index = representative;
-            if constexpr (Dimension >= 2) {
-                if ((column & 1U) != 0) {
-                    index ^= rotation.orbit_masks[0];
-                }
-            }
-            if constexpr (Dimension == 4) {
-                if ((column & 2U) != 0) {
-                    index ^= rotation.orbit_masks[1];
-                }
-            }
-            indices[column] = index;
-            input_real[column] = real[index];
-            input_imag[column] = imag[index];
-        }
-
-        for (size_t row = 0; row < Dimension; ++row) {
-            double output_real = 0.0;
-            double output_imag = 0.0;
-            for (size_t column = 0; column < Dimension; ++column) {
-                const std::complex<double> weight = matrix[row * Dimension + column];
-                output_real +=
-                    weight.real() * input_real[column] - weight.imag() * input_imag[column];
-                output_imag +=
-                    weight.real() * input_imag[column] + weight.imag() * input_real[column];
-            }
-            real[indices[row]] = output_real;
-            imag[indices[row]] = output_imag;
-        }
-    }
-}
-
 uint64_t diagonal_source(const PreparedMeasurement& measurement, uint64_t packed, bool branch) {
     const uint64_t without_pivot = insert_zero_bit(packed, measurement.pivot);
     const bool other_parity =
@@ -272,25 +207,6 @@ void apply_rotation(State& state, const PreparedRotation& rotation, bool sign) n
         apply_nondiagonal_rotation<true>(state, rotation, sine);
     } else {
         apply_nondiagonal_rotation<false>(state, rotation, sine);
-    }
-}
-
-void apply_fused_rotation(State& state, const PreparedFusedRotation& rotation) noexcept {
-    assert(state.active_width() == rotation.active_width &&
-           "fused rotation width must match the active state");
-    switch (rotation.orbit_rank) {
-        case 0:
-            apply_fused_rotation_orbits<1>(state, rotation);
-            return;
-        case 1:
-            apply_fused_rotation_orbits<2>(state, rotation);
-            return;
-        case 2:
-            apply_fused_rotation_orbits<4>(state, rotation);
-            return;
-        default:
-            assert(false && "fused rotation orbit rank must be at most two");
-            return;
     }
 }
