+++ src/clifft/sampling/kernels.h
@@ -3,10 +3,8 @@
 #include "clifft/sampling/plan.h"
 #include "clifft/sampling/state.h"
 
-#include <array>
 #include <complex>
 #include <cstdint>
-#include <vector>
 
 namespace clifft::sampling {
 
@@ -28,20 +26,6 @@ struct PreparedRotation {
     PreparedPauli pauli;
     double cosine = 1.0;
     double sine = 0.0;
-};
-
-// Applies one precomposed unitary to every orbit generated by at most two X
-// masks. Matrices are selected by representative-only Z parities, so each
-// coefficient is still loaded and stored exactly once.
-struct PreparedFusedRotation {
-    uint32_t active_width = 0;
-    uint32_t orbit_rank = 0;
-    // The orbit and selector masks are reduced GF(2) bases in ascending pivot
-    // order; orbit_pivots identifies the bits omitted from representatives.
-    std::array<uint64_t, 2> orbit_masks{};
-    std::array<uint32_t, 2> orbit_pivots{};
-    std::vector<uint64_t> selector_masks;
-    std::vector<std::complex<double>> matrices;
 };
 
 struct PreparedPromotion {
@@ -78,7 +62,6 @@ struct MeasurementProbabilities {
 // Runtime signs have already been evaluated from the plan expression. A true
 // sign negates the Pauli, equivalently negating the prepared sine.
 void apply_rotation(State& state, const PreparedRotation& rotation, bool sign) noexcept;
-void apply_fused_rotation(State& state, const PreparedFusedRotation& rotation) noexcept;
 void apply_promotion(State& state, const PreparedPromotion& promotion, bool sign) noexcept;
 
 // Probability evaluation uses the normalized coefficient arrays and does not
