+++ tests/CMakeLists.txt
@@ -21,6 +21,7 @@ add_executable(clifft_tests
     test_sampling_plan.cc
     test_sampling_planner.cc
     test_sampling_executor.cc
+    test_fused_rotation.cc
     test_sampling_kernels.cc
     test_svm.cc
     test_state_reset.cc
