+++ tests/test_fused_rotation.cc
@@ -1,0 +1,121 @@
+#include "clifft/sampling/executor.h"
+#include "clifft/sampling/kernels.h"
+
+#include <array>
+#include <catch2/catch_test_macros.hpp>
+#include <catch2/matchers/catch_matchers_floating_point.hpp>
+#include <cstddef>
+#include <cstdint>
+#include <optional>
+#include <span>
+#include <variant>
+
+namespace {
+
+using clifft::sampling::ActivePauli;
+using clifft::sampling::AffineBool;
+using clifft::sampling::apply_rotation;
+using clifft::sampling::ExecutablePlan;
+using clifft::sampling::Executor;
+using clifft::sampling::index;
+using clifft::sampling::PlannedAction;
+using clifft::sampling::prepare_rotation;
+using clifft::sampling::RotateActivePauli;
+using clifft::sampling::SamplingPlan;
+using clifft::sampling::State;
+using clifft::sampling::SymbolId;
+using clifft::sampling::SymbolInfo;
+using clifft::sampling::SymbolKind;
+
+SamplingPlan rotation_plan(uint32_t active_width, std::span<const RotateActivePauli> rotations) {
+    SamplingPlan plan;
+    plan.num_qubits = active_width;
+    plan.initial_active_width = active_width;
+    plan.max_active_width = active_width;
+    plan.symbols = {SymbolInfo{SymbolKind::Presampled, std::nullopt, std::nullopt}};
+    for (const RotateActivePauli& rotation : rotations) {
+        plan.actions.push_back(PlannedAction{active_width, active_width, rotation});
+    }
+    return plan;
+}
+
+void require_matches_scalar(const SamplingPlan& plan, uint8_t presampled_value,
+                            size_t expected_action_count) {
+    const ExecutablePlan executable(plan);
+    REQUIRE(executable.num_actions() == expected_action_count);
+    Executor executor(executable);
+    executor.run_shot(std::array<uint8_t, 1>{presampled_value});
+
+    State expected(plan.max_active_width, plan.initial_active_width, plan.global_weight);
+    for (const PlannedAction& planned : plan.actions) {
+        const auto& rotation = std::get<RotateActivePauli>(planned.action);
+        bool sign = rotation.sign.constant();
+        for (SymbolId term : rotation.sign.terms()) {
+            REQUIRE(index(term) == 0);
+            sign ^= presampled_value != 0;
+        }
+        apply_rotation(expected,
+                       prepare_rotation(rotation.pauli, planned.active_before, rotation.half_turns),
+                       sign);
+    }
+
+    REQUIRE(executor.state().size() == expected.size());
+    REQUIRE_THAT(executor.state().global_scalar().real(),
+                 Catch::Matchers::WithinAbs(expected.global_scalar().real(), 1e-12));
+    REQUIRE_THAT(executor.state().global_scalar().imag(),
+                 Catch::Matchers::WithinAbs(expected.global_scalar().imag(), 1e-12));
+    for (uint64_t basis = 0; basis < expected.size(); ++basis) {
+        CAPTURE(presampled_value, basis);
+        REQUIRE_THAT(executor.state().real_data()[basis],
+                     Catch::Matchers::WithinAbs(expected.real_data()[basis], 1e-12));
+        REQUIRE_THAT(executor.state().imag_data()[basis],
+                     Catch::Matchers::WithinAbs(expected.imag_data()[basis], 1e-12));
+    }
+}
+
+}  // namespace
+
+TEST_CASE("Fused rotation stops at a dynamic sign") {
+    const std::array rotations = {
+        RotateActivePauli{{0b01, 0b00}, 0.25, AffineBool(false)},
+        RotateActivePauli{{0b10, 0b01}, -0.3, AffineBool(true)},
+        RotateActivePauli{{0b11, 0b11}, 0.4, AffineBool(false)},
+        RotateActivePauli{{0b00, 0b10}, 0.2, AffineBool::symbol(SymbolId{0})},
+        RotateActivePauli{{0b01, 0b10}, -0.1, AffineBool(false)},
+        RotateActivePauli{{0b10, 0b11}, 0.35, AffineBool(true)},
+        RotateActivePauli{{0b11, 0b01}, -0.45, AffineBool(false)},
+    };
+    const SamplingPlan plan = rotation_plan(2, rotations);
+    require_matches_scalar(plan, 0, 3);
+    require_matches_scalar(plan, 1, 3);
+}
+
+TEST_CASE("Fused rotation preserves a signed identity barrier") {
+    const std::array rotations = {
+        RotateActivePauli{{0b01, 0b00}, 0.25, AffineBool(false)},
+        RotateActivePauli{{0b10, 0b01}, -0.3, AffineBool(true)},
+        RotateActivePauli{{0b11, 0b11}, 0.4, AffineBool(false)},
+        RotateActivePauli{{0b00, 0b00}, 0.5, AffineBool(true)},
+        RotateActivePauli{{0b01, 0b10}, -0.1, AffineBool(false)},
+        RotateActivePauli{{0b10, 0b11}, 0.35, AffineBool(true)},
+        RotateActivePauli{{0b11, 0b01}, -0.45, AffineBool(false)},
+    };
+    require_matches_scalar(rotation_plan(2, rotations), 0, 3);
+}
+
+TEST_CASE("Fused rotation leaves a below threshold run unfused") {
+    const std::array rotations = {
+        RotateActivePauli{{0b01, 0b10}, 0.25, AffineBool(false)},
+        RotateActivePauli{{0b10, 0b01}, -0.3, AffineBool(true)},
+    };
+    require_matches_scalar(rotation_plan(2, rotations), 0, 2);
+}
+
+TEST_CASE("Fused rotation falls back for a rank three run") {
+    const std::array rotations = {
+        RotateActivePauli{{0b001, 0b010}, 0.25, AffineBool(false)},
+        RotateActivePauli{{0b010, 0b100}, -0.3, AffineBool(true)},
+        RotateActivePauli{{0b100, 0b001}, 0.4, AffineBool(false)},
+    };
+    require_matches_scalar(rotation_plan(3, rotations), 0, 3);
+}
